Extract System Details sliver into new widget
